Add format code button to nav
