+++ app/(navigation)/(code)/code.tsx
@@ -21,6 +21,9 @@ import tailwindDark from "./assets/tailwind/dark.json";
 import ExportButton from "./components/ExportButton";
 import { NavigationActions } from "@/components/navigation";
 import { InfoDialog } from "./components/InfoDialog";
+import { WandIcon } from "@raycast/icons";
+import { Button } from "@/components/button";
+import FormatButton from "./components/FormatCodeButton";
 
 export function Code() {
   const [highlighter, setHighlighter] = useAtom(highlighterAtom);
@@ -40,6 +43,7 @@ export function Code() {
       <FrameContextStore>
         <NavigationActions>
           <InfoDialog />
+          <FormatButton />
           <ExportButton />
         </NavigationActions>
         <div className={styles.app}>
